Remove the comp_dir_substitutions argument from package_source_files_from_rr_output.
With --gdb-script rr can synthesize comp dir substitutions from the output of the gdb
script, so gating the substitution logic on the user having passed in a substitution
on the command line is no longer accurate. With that removed, there's no point to
passing comp_dir_substitutions to package_source_files_from_rr_output at all.